Migrate collectd_dsl function to new puppet function interface
